Only load update info modal the first time after an update

+++ src/bp-core/classes/class-bp-admin.php
@@ -1129,6 +1129,10 @@ if ( ! class_exists( 'BP_Admin' ) ) :
 		 */
 		public function bb_display_update_plugin_information() {
 			// Check the transient to see if we've just updated the plugin.
+			if ( ! get_option( '_bb_is_update' ) ) {
+				return;
+			}
+
 			global $bp;
 			include trailingslashit( $bp->plugin_dir . 'bp-core/admin' ) . 'templates/update-buddyboss.php';
 			delete_option( '_bb_is_update' );
